improved performance of escape modifer on escaping html, htmlall, url, urlpathinfo, quotes, javascript.
Thereby introducing a new pattern to modifier development:

For the given escape routines the compiler function determines optimization possibility. 
If optimization is possible, the program-code executed by the regular escape is written to the compiled template.
If optimization can't be done (say because an argument is a variable instead of a literal) the compiler function generates the same compiled template output as the regular modifier would have.